+++ net/minecraft/world/entity/animal/MushroomCow.java
@@ -179,7 +179,7 @@ implements Shearable {
         super.addAdditionalSaveData(compoundTag);
         compoundTag.putString("Type", this.getMushroomType().type);
         if (this.effect != null) {
-            compoundTag.putByte("EffectId", (byte)MobEffect.getId(this.effect));
+            compoundTag.putInt("EffectId", MobEffect.getId(this.effect));
             compoundTag.putInt("EffectDuration", this.effectDuration);
         }
     }
@@ -189,7 +189,7 @@ implements Shearable {
         super.readAdditionalSaveData(compoundTag);
         this.setMushroomType(MushroomType.byType(compoundTag.getString("Type")));
         if (compoundTag.contains("EffectId", 1)) {
-            this.effect = MobEffect.byId(compoundTag.getByte("EffectId"));
+            this.effect = MobEffect.byId(compoundTag.getInt("EffectId"));
         }
         if (compoundTag.contains("EffectDuration", 3)) {
             this.effectDuration = compoundTag.getInt("EffectDuration");
